Split node command up into multiple files
