+++ peer/node/node.go
@@ -17,37 +17,9 @@ limitations under the License.
 package node
 
 import (
-	"bytes"
-	"errors"
 	"fmt"
-	"google/protobuf"
-	"io/ioutil"
-	"net"
-	"net/http"
-	"os"
-	"os/signal"
-	"path/filepath"
-	"strconv"
-	"sync"
-	"syscall"
-	"time"
 
-	"golang.org/x/net/context"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials"
-	"google.golang.org/grpc/grpclog"
-
-	"github.com/hyperledger/fabric/consensus/helper"
 	"github.com/hyperledger/fabric/core"
-	"github.com/hyperledger/fabric/core/chaincode"
-	"github.com/hyperledger/fabric/core/comm"
-	"github.com/hyperledger/fabric/core/crypto"
-	"github.com/hyperledger/fabric/core/ledger/genesis"
-	"github.com/hyperledger/fabric/core/peer"
-	"github.com/hyperledger/fabric/core/rest"
-	"github.com/hyperledger/fabric/core/system_chaincode"
-	"github.com/hyperledger/fabric/events/producer"
-	pb "github.com/hyperledger/fabric/protos"
 	"github.com/op/go-logging"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
@@ -56,21 +28,17 @@ import (
 const nodeFuncName = "node"
 
 var (
-	stopPidFile      string
-	chaincodeDevMode bool
+	stopPidFile string
 )
 var logger = logging.MustGetLogger("nodeCmd")
 
+// Cmd returns the cobra command for Node
 func Cmd() *cobra.Command {
-	nodeCmd.AddCommand(nodeStartCmd)
+	nodeCmd.AddCommand(startCmd())
 	nodeCmd.AddCommand(nodeStatusCmd)
 
 	nodeStopCmd.Flags().StringVar(&stopPidFile, "stop-peer-pid-file", viper.GetString("peer.fileSystemPath"), "Location of peer pid local file, for forces kill")
 	nodeCmd.AddCommand(nodeStopCmd)
-
-	// Set the flags on the node start command.
-	flags := nodeStartCmd.Flags()
-	flags.BoolVarP(&chaincodeDevMode, "peer-chaincodedev", "", false, "Whether peer in chaincode development mode")
 
 	return nodeCmd
 }
@@ -82,412 +50,4 @@ var nodeCmd = &cobra.Command{
 	PersistentPreRun: func(cmd *cobra.Command, args []string) {
 		core.LoggingInit(nodeFuncName)
 	},
-}
-
-var nodeStartCmd = &cobra.Command{
-	Use:   "start",
-	Short: "Starts the node.",
-	Long:  `Starts a node that interacts with the network.`,
-	RunE: func(cmd *cobra.Command, args []string) error {
-		return serve(args)
-	},
-}
-
-var nodeStatusCmd = &cobra.Command{
-	Use:   "status",
-	Short: "Returns status of the node.",
-	Long:  `Returns the status of the running node.`,
-	Run: func(cmd *cobra.Command, args []string) {
-		status()
-	},
-}
-
-var nodeStopCmd = &cobra.Command{
-	Use:   "stop",
-	Short: "Stops the running node.",
-	Long:  `Stops the running node, disconnecting from the network.`,
-	Run: func(cmd *cobra.Command, args []string) {
-		stop()
-	},
-}
-
-func serve(args []string) error {
-	// Parameter overrides must be processed before any paramaters are
-	// cached. Failures to cache cause the server to terminate immediately.
-	if chaincodeDevMode {
-		logger.Info("Running in chaincode development mode")
-		logger.Info("Set consensus to NOOPS and user starts chaincode")
-		logger.Info("Disable loading validity system chaincode")
-
-		viper.Set("peer.validator.enabled", "true")
-		viper.Set("peer.validator.consensus", "noops")
-		viper.Set("chaincode.mode", chaincode.DevModeUserRunsChaincode)
-
-	}
-
-	if err := peer.CacheConfiguration(); err != nil {
-		return err
-	}
-
-	peerEndpoint, err := peer.GetPeerEndpoint()
-	if err != nil {
-		err = fmt.Errorf("Failed to get Peer Endpoint: %s", err)
-		return err
-	}
-
-	listenAddr := viper.GetString("peer.listenAddress")
-
-	if "" == listenAddr {
-		logger.Debug("Listen address not specified, using peer endpoint address")
-		listenAddr = peerEndpoint.Address
-	}
-
-	lis, err := net.Listen("tcp", listenAddr)
-	if err != nil {
-		grpclog.Fatalf("Failed to listen: %v", err)
-	}
-
-	ehubLis, ehubGrpcServer, err := createEventHubServer()
-	if err != nil {
-		grpclog.Fatalf("Failed to create ehub server: %v", err)
-	}
-
-	logger.Infof("Security enabled status: %t", core.SecurityEnabled())
-	if viper.GetBool("security.privacy") {
-		if core.SecurityEnabled() {
-			logger.Infof("Privacy enabled status: true")
-		} else {
-			panic(errors.New("Privacy cannot be enabled as requested because security is disabled"))
-		}
-	} else {
-		logger.Infof("Privacy enabled status: false")
-	}
-
-	var opts []grpc.ServerOption
-	if comm.TLSEnabled() {
-		creds, err := credentials.NewServerTLSFromFile(viper.GetString("peer.tls.cert.file"), viper.GetString("peer.tls.key.file"))
-		if err != nil {
-			grpclog.Fatalf("Failed to generate credentials %v", err)
-		}
-		opts = []grpc.ServerOption{grpc.Creds(creds)}
-	}
-
-	grpcServer := grpc.NewServer(opts...)
-
-	secHelper, err := getSecHelper()
-	if err != nil {
-		return err
-	}
-
-	secHelperFunc := func() crypto.Peer {
-		return secHelper
-	}
-
-	registerChaincodeSupport(chaincode.DefaultChain, grpcServer, secHelper)
-
-	var peerServer *peer.PeerImpl
-
-	// Create the peerServer
-	if peer.ValidatorEnabled() {
-		logger.Debug("Running as validating peer - making genesis block if needed")
-		makeGenesisError := genesis.MakeGenesis()
-		if makeGenesisError != nil {
-			return makeGenesisError
-		}
-		logger.Debugf("Running as validating peer - installing consensus %s", viper.GetString("peer.validator.consensus"))
-		peerServer, err = peer.NewPeerWithEngine(secHelperFunc, helper.GetEngine)
-	} else {
-		logger.Debug("Running as non-validating peer")
-		peerServer, err = peer.NewPeerWithHandler(secHelperFunc, peer.NewPeerHandler)
-	}
-
-	if err != nil {
-		logger.Fatalf("Failed creating new peer with handler %v", err)
-
-		return err
-	}
-
-	// Register the Peer server
-	pb.RegisterPeerServer(grpcServer, peerServer)
-
-	// Register the Admin server
-	pb.RegisterAdminServer(grpcServer, core.NewAdminServer())
-
-	// Register Devops server
-	serverDevops := core.NewDevopsServer(peerServer)
-	pb.RegisterDevopsServer(grpcServer, serverDevops)
-
-	// Register the ServerOpenchain server
-	serverOpenchain, err := rest.NewOpenchainServerWithPeerInfo(peerServer)
-	if err != nil {
-		err = fmt.Errorf("Error creating OpenchainServer: %s", err)
-		return err
-	}
-
-	pb.RegisterOpenchainServer(grpcServer, serverOpenchain)
-
-	// Create and register the REST service if configured
-	if viper.GetBool("rest.enabled") {
-		go rest.StartOpenchainRESTServer(serverOpenchain, serverDevops)
-	}
-
-	logger.Infof("Starting peer with ID=%s, network ID=%s, address=%s, rootnodes=%v, validator=%v",
-		peerEndpoint.ID, viper.GetString("peer.networkId"), peerEndpoint.Address, viper.GetString("peer.discovery.rootnode"), peer.ValidatorEnabled())
-
-	// Start the grpc server. Done in a goroutine so we can deploy the
-	// genesis block if needed.
-	serve := make(chan error)
-
-	sigs := make(chan os.Signal, 1)
-	signal.Notify(sigs, syscall.SIGINT, syscall.SIGTERM)
-	go func() {
-		sig := <-sigs
-		fmt.Println()
-		fmt.Println(sig)
-		serve <- nil
-	}()
-
-	go func() {
-		var grpcErr error
-		if grpcErr = grpcServer.Serve(lis); grpcErr != nil {
-			grpcErr = fmt.Errorf("grpc server exited with error: %s", grpcErr)
-		} else {
-			logger.Info("grpc server exited")
-		}
-		serve <- grpcErr
-	}()
-
-	if err := writePid(viper.GetString("peer.fileSystemPath")+"/peer.pid", os.Getpid()); err != nil {
-		return err
-	}
-
-	// Start the event hub server
-	if ehubGrpcServer != nil && ehubLis != nil {
-		go ehubGrpcServer.Serve(ehubLis)
-	}
-
-	if viper.GetBool("peer.profile.enabled") {
-		go func() {
-			profileListenAddress := viper.GetString("peer.profile.listenAddress")
-			logger.Infof("Starting profiling server with listenAddress = %s", profileListenAddress)
-			if profileErr := http.ListenAndServe(profileListenAddress, nil); profileErr != nil {
-				logger.Errorf("Error starting profiler: %s", profileErr)
-			}
-		}()
-	}
-
-	// Block until grpc server exits
-	return <-serve
-}
-
-func status() (err error) {
-	clientConn, err := peer.NewPeerClientConnection()
-	if err != nil {
-		logger.Infof("Error trying to connect to local peer: %s", err)
-		err = fmt.Errorf("Error trying to connect to local peer: %s", err)
-		fmt.Println(&pb.ServerStatus{Status: pb.ServerStatus_UNKNOWN})
-		return err
-	}
-
-	serverClient := pb.NewAdminClient(clientConn)
-
-	status, err := serverClient.GetStatus(context.Background(), &google_protobuf.Empty{})
-	if err != nil {
-		logger.Infof("Error trying to get status from local peer: %s", err)
-		err = fmt.Errorf("Error trying to connect to local peer: %s", err)
-		fmt.Println(&pb.ServerStatus{Status: pb.ServerStatus_UNKNOWN})
-		return err
-	}
-	fmt.Println(status)
-	return nil
-}
-
-func stop() (err error) {
-	clientConn, err := peer.NewPeerClientConnection()
-	if err != nil {
-		pidFile := stopPidFile + "/peer.pid"
-		//fmt.Printf("Stopping local peer using process pid from %s \n", pidFile)
-		logger.Infof("Error trying to connect to local peer: %s", err)
-		logger.Infof("Stopping local peer using process pid from %s", pidFile)
-		pid, ferr := readPid(pidFile)
-		if ferr != nil {
-			err = fmt.Errorf("Error trying to read pid from %s: %s", pidFile, ferr)
-			return
-		}
-		killerr := syscall.Kill(pid, syscall.SIGTERM)
-		if killerr != nil {
-			err = fmt.Errorf("Error trying to kill -9 pid %d: %s", pid, killerr)
-			return
-		}
-		return nil
-	}
-	logger.Info("Stopping peer using grpc")
-	serverClient := pb.NewAdminClient(clientConn)
-
-	status, err := serverClient.StopServer(context.Background(), &google_protobuf.Empty{})
-	if err != nil {
-		fmt.Println(&pb.ServerStatus{Status: pb.ServerStatus_STOPPED})
-		return nil
-	}
-
-	err = fmt.Errorf("Connection remain opened, peer process doesn't exit")
-	fmt.Println(status)
-	return err
-}
-
-func createEventHubServer() (net.Listener, *grpc.Server, error) {
-	var lis net.Listener
-	var grpcServer *grpc.Server
-	var err error
-	if peer.ValidatorEnabled() {
-		lis, err = net.Listen("tcp", viper.GetString("peer.validator.events.address"))
-		if err != nil {
-			return nil, nil, fmt.Errorf("failed to listen: %v", err)
-		}
-
-		//TODO - do we need different SSL material for events ?
-		var opts []grpc.ServerOption
-		if comm.TLSEnabled() {
-			creds, err := credentials.NewServerTLSFromFile(viper.GetString("peer.tls.cert.file"), viper.GetString("peer.tls.key.file"))
-			if err != nil {
-				return nil, nil, fmt.Errorf("Failed to generate credentials %v", err)
-			}
-			opts = []grpc.ServerOption{grpc.Creds(creds)}
-		}
-
-		grpcServer = grpc.NewServer(opts...)
-		ehServer := producer.NewEventsServer(uint(viper.GetInt("peer.validator.events.buffersize")), viper.GetInt("peer.validator.events.timeout"))
-		pb.RegisterEventsServer(grpcServer, ehServer)
-	}
-	return lis, grpcServer, err
-}
-
-var once sync.Once
-
-//this should be called exactly once and the result cached
-//NOTE- this crypto func might rightly belong in a crypto package
-//and universally accessed
-func getSecHelper() (crypto.Peer, error) {
-	var secHelper crypto.Peer
-	var err error
-	once.Do(func() {
-		if core.SecurityEnabled() {
-			enrollID := viper.GetString("security.enrollID")
-			enrollSecret := viper.GetString("security.enrollSecret")
-			if peer.ValidatorEnabled() {
-				logger.Debugf("Registering validator with enroll ID: %s", enrollID)
-				if err = crypto.RegisterValidator(enrollID, nil, enrollID, enrollSecret); nil != err {
-					return
-				}
-				logger.Debugf("Initializing validator with enroll ID: %s", enrollID)
-				secHelper, err = crypto.InitValidator(enrollID, nil)
-				if nil != err {
-					return
-				}
-			} else {
-				logger.Debugf("Registering non-validator with enroll ID: %s", enrollID)
-				if err = crypto.RegisterPeer(enrollID, nil, enrollID, enrollSecret); nil != err {
-					return
-				}
-				logger.Debugf("Initializing non-validator with enroll ID: %s", enrollID)
-				secHelper, err = crypto.InitPeer(enrollID, nil)
-				if nil != err {
-					return
-				}
-			}
-		}
-	})
-	return secHelper, err
-}
-
-func registerChaincodeSupport(chainname chaincode.ChainName, grpcServer *grpc.Server, secHelper crypto.Peer) {
-	//get user mode
-	userRunsCC := false
-	if viper.GetString("chaincode.mode") == chaincode.DevModeUserRunsChaincode {
-		userRunsCC = true
-	}
-
-	//get chaincode startup timeout
-	tOut, err := strconv.Atoi(viper.GetString("chaincode.startuptimeout"))
-	if err != nil { //what went wrong ?
-		fmt.Printf("could not retrive timeout var...setting to 5secs\n")
-		tOut = 5000
-	}
-	ccStartupTimeout := time.Duration(tOut) * time.Millisecond
-
-	ccSrv := chaincode.NewChaincodeSupport(chainname, peer.GetPeerEndpoint, userRunsCC, ccStartupTimeout, secHelper)
-
-	//Now that chaincode is initialized, register all system chaincodes.
-	system_chaincode.RegisterSysCCs()
-
-	pb.RegisterChaincodeSupportServer(grpcServer, ccSrv)
-}
-
-func writePid(fileName string, pid int) error {
-	err := os.MkdirAll(filepath.Dir(fileName), 0755)
-	if err != nil {
-		return err
-	}
-
-	fd, err := os.OpenFile(fileName, os.O_RDWR|os.O_CREATE, 0644)
-	if err != nil {
-		return err
-	}
-	defer fd.Close()
-	if err := syscall.Flock(int(fd.Fd()), syscall.LOCK_EX|syscall.LOCK_NB); err != nil {
-		return fmt.Errorf("can't lock '%s', lock is held", fd.Name())
-	}
-
-	if _, err := fd.Seek(0, 0); err != nil {
-		return err
-	}
-
-	if err := fd.Truncate(0); err != nil {
-		return err
-	}
-
-	if _, err := fmt.Fprintf(fd, "%d", pid); err != nil {
-		return err
-	}
-
-	if err := fd.Sync(); err != nil {
-		return err
-	}
-
-	if err := syscall.Flock(int(fd.Fd()), syscall.LOCK_UN); err != nil {
-		return fmt.Errorf("can't release lock '%s', lock is held", fd.Name())
-	}
-	return nil
-}
-
-func readPid(fileName string) (int, error) {
-	fd, err := os.OpenFile(fileName, os.O_RDWR|os.O_CREATE, 0644)
-	if err != nil {
-		return 0, err
-	}
-	defer fd.Close()
-	if err := syscall.Flock(int(fd.Fd()), syscall.LOCK_EX|syscall.LOCK_NB); err != nil {
-		return 0, fmt.Errorf("can't lock '%s', lock is held", fd.Name())
-	}
-
-	if _, err := fd.Seek(0, 0); err != nil {
-		return 0, err
-	}
-
-	data, err := ioutil.ReadAll(fd)
-	if err != nil {
-		return 0, err
-	}
-
-	pid, err := strconv.Atoi(string(bytes.TrimSpace(data)))
-	if err != nil {
-		return 0, fmt.Errorf("error parsing pid from %s: %s", fd.Name(), err)
-	}
-
-	if err := syscall.Flock(int(fd.Fd()), syscall.LOCK_UN); err != nil {
-		return 0, fmt.Errorf("can't release lock '%s', lock is held", fd.Name())
-	}
-
-	return pid, nil
-
 }
